backend/requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
azure-storage-blob
azure-ai-textanalytics
azure-search-documents
azure-cognitiveservices-speech
azure-communication-email
openai>=1.30.0
langchain>=0.3.0
langchain-openai>=0.2.0
psycopg2-binary
pyjwt[crypto]
stripe>=8.0.0
mercadopago
python-dotenv
requests
beautifulsoup4
playwright
applicationinsights
pytest
pytest-cov
pytest-asyncio
